refractored Route and Node classes to correct location in directories
