+++ src/main/java/memoranda/Node.java
@@ -1,0 +1,32 @@
+package main.java.memoranda;
+
+public class Node {
+    double latitude;
+    double longitude;
+
+    Node(double lat, double lon){
+        latitude = lat;
+        longitude = lon;
+    }
+
+    public double getLatitude(){
+        return latitude;
+    }
+    public double getLongitude(){
+        return longitude;
+    }
+
+    public static double distanceOfNodes(Node n, Node nn){
+        double R = 6371; // radius of the earth in km;
+        double dLat = Math.toRadians(n.getLatitude() - nn.getLatitude());
+        double dLon = Math.toRadians(n.getLongitude() - nn.getLongitude());
+
+        double a = Math.sin(dLat/2) * Math.sin(dLat/2) +
+                Math.cos(Math.toRadians(nn.getLatitude())) * Math.cos(Math.toRadians(n.getLatitude())) *
+                        Math.sin(dLon/2) * Math.sin(dLon/2);
+        double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
+        double d = R * c; //distance in kilometers
+        return d;
+    }
+
+}
