+++ src/main/java/memoranda/util/Util.java
@@ -18,6 +18,7 @@ import java.util.Iterator;
 
 import javax.swing.JFileChooser;
 
+import main.java.memoranda.Node;
 import main.java.memoranda.date.CalendarDate;
 import main.java.memoranda.ui.App;
 import main.java.memoranda.ui.AppFrame;
@@ -181,18 +182,5 @@ public class Util {
             stackTrace = ste[i].toString() + "\n";
         }
         System.out.println("[ERROR] Stack Trace: " + stackTrace);
-    }
-
-    public static double distanceOfNodes(Node n, Node nn){
-        double R = 6371; // radius of the earth in km;
-        double dLat = Math.toRadians(n.getLatitude() - nn.getLatitude());
-        double dLon = Math.toRadians(n.getLongitude() - nn.getLongitude());
-
-        double a = Math.sin(dLat/2) * Math.sin(dLat/2) +
-                Math.cos(Math.toRadians(nn.getLatitude())) * Math.cos(Math.toRadians(n.getLatitude())) *
-                        Math.sin(dLon/2) * Math.sin(dLon/2);
-        double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-        double d = R * c; //distance in kilometers
-        return d;
     }
 }
